pyproject.toml
[project]
name = "aragsys"
version = "0.1.0"
description = "Scientific Agentic RAG Framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.0",
    "chromadb>=1.4.1",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.0",
    "lxml>=5.3.0",
    "neo4j>=5.0.0",         # Neo4j Python driver
    "pdfplumber>=0.11.4",
    "pocketflow>=0.0.3",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "spacy>=3.8.0",         # NER for entity extraction
    "uvicorn>=0.32.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
]

[dependency-groups]
test = []
